sign batch
