+++ contracts/multisig/interact/src/multisig_interact.rs
@@ -7,7 +7,7 @@ mod multisig_interact_wegld;
 use clap::Parser;
 use multisig::{
     multisig_perform::ProxyTrait as _, multisig_propose::ProxyTrait as _,
-    multisig_state::ProxyTrait as _, ProxyTrait as _,
+    multisig_sign::ProxyTrait, multisig_state::ProxyTrait as _, ProxyTrait as _,
 };
 use multisig_interact_config::Config;
 use multisig_interact_state::State;
@@ -43,52 +43,52 @@ async fn main() {
     match &cli.command {
         Some(multisig_interact_cli::InteractCliCommand::Board) => {
             multisig_interact.print_board().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::Deploy) => {
             multisig_interact.deploy().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::DnsRegister(args)) => {
             multisig_interact.dns_register(&args.name).await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::Feed) => {
             multisig_interact.feed_contract_egld().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::MultiDeploy(args)) => {
             multisig_interact.multi_deploy(&args.count).await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::NftFullAllRoles) => {
             multisig_interact
                 .issue_multisig_and_collection_with_all_roles_full()
                 .await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::NftFull) => {
             multisig_interact.issue_multisig_and_collection_full().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::NftIssueAllRoles) => {
             multisig_interact.issue_collection_with_all_roles().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::NftIssue) => {
             multisig_interact.issue_collection().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::NftItems) => {
             multisig_interact.create_items().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::NftSpecial) => {
             multisig_interact.set_special_role().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::Quorum) => {
             multisig_interact.print_quorum().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::UnwrapEgld) => {
             multisig_interact.unwrap_egld().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::WEgldSwapFull) => {
             multisig_interact.wegld_swap_full().await;
-        },
+        }
         Some(multisig_interact_cli::InteractCliCommand::WrapEgld) => {
             multisig_interact.wrap_egld().await;
-        },
-        None => {},
+        }
+        None => {}
     }
 }
