+++ contracts/multisig/scenarios/call_other_shard-1.scen.json
@@ -34,7 +34,8 @@
                 "function": "proposeTransferExecute",
                 "arguments": [
                     "sc:other-shard-1",
-                    "10"
+                    "10",
+                    "0x"
                 ],
                 "gasLimit": "100,000,000",
                 "gasPrice": "0"
