+++ contracts/multisig/scenarios/call_other_shard-2.scen.json
@@ -27,6 +27,7 @@
                 "arguments": [
                     "sc:other-shard-2",
                     "0",
+                    "0x",
                     "str:method-from-other-shard",
                     "str:arg1",
                     "str:arg2"
