+++ contracts/multisig/src/multisig.rs
@@ -4,6 +4,7 @@ pub mod action;
 pub mod multisig_events;
 pub mod multisig_perform;
 pub mod multisig_propose;
+pub mod multisig_sign;
 pub mod multisig_state;
 pub mod user_role;
 
@@ -19,6 +20,7 @@ multiversx_sc::imports!();
 pub trait Multisig:
     multisig_state::MultisigStateModule
     + multisig_propose::MultisigProposeModule
+    + multisig_sign::MultisigSignModule
     + multisig_perform::MultisigPerformModule
     + multisig_events::MultisigEventsModule
     + multiversx_sc_modules::dns::DnsModule
@@ -75,18 +77,6 @@ pub trait Multisig:
         result
     }
 
-    /// Returns `true` (`1`) if the user has signed the action.
-    /// Does not check whether or not the user is still a board member and the signature valid.
-    #[view]
-    fn signed(&self, user: ManagedAddress, action_id: usize) -> bool {
-        let user_id = self.user_mapper().get_user_id(&user);
-        if user_id == 0 {
-            false
-        } else {
-            self.action_signer_ids(action_id).contains(&user_id)
-        }
-    }
-
     /// Indicates user rights.
     /// `0` = no rights,
     /// `1` = can propose, but not sign,
@@ -127,37 +117,6 @@ pub trait Multisig:
             }
         }
         result
-    }
-
-    /// Used by board members to sign actions.
-    #[endpoint]
-    fn sign(&self, action_id: usize) {
-        require!(
-            !self.action_mapper().item_is_empty_unchecked(action_id),
-            "action does not exist"
-        );
-
-        let (caller_id, caller_role) = self.get_caller_id_and_role();
-        require!(caller_role.can_sign(), "only board members can sign");
-
-        if !self.action_signer_ids(action_id).contains(&caller_id) {
-            self.action_signer_ids(action_id).insert(caller_id);
-        }
-    }
-
-    /// Board members can withdraw their signatures if they no longer desire for the action to be executed.
-    /// Actions that are left with no valid signatures can be then deleted to free up storage.
-    #[endpoint]
-    fn unsign(&self, action_id: usize) {
-        require!(
-            !self.action_mapper().item_is_empty_unchecked(action_id),
-            "action does not exist"
-        );
-
-        let (caller_id, caller_role) = self.get_caller_id_and_role();
-        require!(caller_role.can_sign(), "only board members can un-sign");
-
-        self.action_signer_ids(action_id).swap_remove(&caller_id);
     }
 
     /// Clears storage pertaining to an action that is no longer supposed to be executed.
